server/requirements.txt
flask
flask-cors
pyyaml
google-genai
python-dotenv
